feat: extend adapter registry to indexers + media servers
Open IndexerType/MediaServerType to string, remove DB enum constraints,
split adapters into TorznabAdapter/PlexAdapter, extend AdapterRegistry
with register/get/list for all three adapter kinds, validate types via
registry in services, add listTypes endpoints. Closes #31.
